move CustomerService file.
